Hopefully fixed a game crash
The code opens the new 1.2.13 sub-chunk format and writes it back to the old sub-chunk format.
The bug was that it wasn't changing the format version to the old version so the game was confused. Also added a catch for the malformed version which get assigned to the old sub-chunk version